Refine SDK package structure and move internal classes to runtime
