+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/inbound/correlation/BoundaryEventCorrelationPoint.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.connector.api.inbound.correlation;
+package io.camunda.connector.runtime.core.inbound.correlation;
 
 public record BoundaryEventCorrelationPoint(
     String messageName,
