+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/inbound/correlation/InboundCorrelationHandler.java
@@ -19,10 +19,6 @@ package io.camunda.connector.runtime.core.inbound.correlation;
 import io.camunda.connector.api.error.ConnectorException;
 import io.camunda.connector.api.error.ConnectorInputException;
 import io.camunda.connector.api.inbound.InboundConnectorResult;
-import io.camunda.connector.api.inbound.correlation.BoundaryEventCorrelationPoint;
-import io.camunda.connector.api.inbound.correlation.MessageCorrelationPoint;
-import io.camunda.connector.api.inbound.correlation.MessageStartEventCorrelationPoint;
-import io.camunda.connector.api.inbound.correlation.StartEventCorrelationPoint;
 import io.camunda.connector.api.inbound.result.CorrelatedMessage;
 import io.camunda.connector.api.inbound.result.CorrelatedMessageStart;
 import io.camunda.connector.api.inbound.result.CorrelationErrorData;
