+++ connector-runtime/connector-runtime-core/src/main/java/io/camunda/connector/runtime/core/inbound/correlation/StartEventCorrelationPoint.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.connector.api.inbound.correlation;
+package io.camunda.connector.runtime.core.inbound.correlation;
 
 /** Properties of a StartEvent triggered by an Inbound Connector */
 public record StartEventCorrelationPoint(
