+++ connector-sdk/test/src/main/java/io/camunda/connector/test/inbound/InboundConnectorDefinitionBuilder.java
@@ -17,8 +17,8 @@
 package io.camunda.connector.test.inbound;
 
 import io.camunda.connector.api.inbound.InboundConnectorDefinition;
-import io.camunda.connector.api.inbound.correlation.ProcessCorrelationPoint;
-import io.camunda.connector.api.inbound.correlation.StartEventCorrelationPoint;
+import io.camunda.connector.runtime.core.inbound.correlation.ProcessCorrelationPoint;
+import io.camunda.connector.runtime.core.inbound.correlation.StartEventCorrelationPoint;
 
 /** Test helper class for creating an {@link InboundConnectorDefinition} with a fluent API. */
 public class InboundConnectorDefinitionBuilder {
