+++ connectors/webhook/src/test/java/io/camunda/connector/inbound/HttpWebhookExecutableTest.java
@@ -14,8 +14,8 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import com.google.common.net.HttpHeaders;
 import com.google.common.net.MediaType;
-import io.camunda.connector.api.error.WebhookConnectorException;
 import io.camunda.connector.api.inbound.InboundConnectorContext;
+import io.camunda.connector.api.inbound.webhook.WebhookConnectorException;
 import io.camunda.connector.api.inbound.webhook.WebhookProcessingPayload;
 import io.camunda.connector.inbound.signature.HMACAlgoCustomerChoice;
 import io.camunda.connector.inbound.utils.HttpMethods;
