+++ connectors/webhook/src/test/java/io/camunda/connector/inbound/authorization/JWTIntegrationTest.java
@@ -24,8 +24,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.common.ClasspathFileSource;
-import io.camunda.connector.api.error.WebhookConnectorException;
 import io.camunda.connector.api.inbound.result.MessageCorrelationResult;
+import io.camunda.connector.api.inbound.webhook.WebhookConnectorException;
 import io.camunda.connector.api.inbound.webhook.WebhookResult;
 import io.camunda.connector.feel.ConnectorsObjectMapperSupplier;
 import io.camunda.connector.inbound.HttpWebhookExecutable;
